requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_sysroot_linux-64_curr_repodata_hack=3=h69a702a_16
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.3.4=pypi_0
aiohttp=3.10.0=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
binutils_impl_linux-64=2.40=ha1999f0_7
black=24.8.0=pypi_0
blas=1.0=mkl
boto3=1.34.153=pypi_0
botocore=1.34.153=pypi_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.4.0=pypi_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
click-prompt=0.5.1=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
decorator=4.4.2=pypi_0
detectron2=0.6=dev_0
diff-gaussian-rasterization=0.0.0=dev_0
diffdist=0.1=pypi_0
docker-pycreds=0.4.0=pypi_0
dotmap=1.3.30=pypi_0
einops=0.3.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=h4a9f257_0
freetype-py=2.4.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20210915=py38
gcc=14.1.0=h6f9ffa1_0
gcc_impl_linux-64=14.1.0=h3c94d91_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.32.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.65.1=pypi_0
huggingface-hub=0.24.5=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.7=py38h06a4308_0
imageio=2.34.2=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.2.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.9=py38
jinja2=3.1.4=py38h06a4308_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_2
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
kernel-headers_linux-64=3.10.0=h4a8ded7_16
kiwisolver=1.4.5=pypi_0
kornia=0.6.0=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=hf3520f5_7
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.10.1.7=0
libcurand=10.3.6.82=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-devel_linux-64=14.1.0=h5d3d1c9_100
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libsanitizer=14.1.0=hcba0ae0_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lightning=2.3.3=pypi_0
lightning-utilities=0.11.6=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py38h5eee18b_0
mask2former=0.1=pypi_0
matplotlib=3.7.5=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
moviepy=1.0.3=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
odise=0.1=dev_0
omegaconf=2.1.1=pypi_0
open-clip-torch=2.0.2=pypi_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.3.1=h4bc722e_2
packaging=24.1=pypi_0
pandas=1.4.1=pypi_0
pathspec=0.12.1=pypi_0
pathtools=0.1.2=pypi_0
peract-bimanual=0.0.1=pypi_0
perceiver-pytorch=0.8.8=pypi_0
pillow=9.5.0=pypi_0
pip=24.0=py38h06a4308_0
platformdirs=4.2.2=pypi_0
portalocker=2.10.1=py38h578d9bd_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=4.25.4=pypi_0
psutil=6.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pydeprecate=0.3.1=pypi_0
pyglet=2.0.16=pypi_0
pygments=2.18.0=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
pyre-extensions=0.0.23=pypi_0
pyrender=0.1.45=pypi_0
pyrep=4.1.0.3=dev_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-lightning=1.4.2=pypi_0
pytorch-mutex=1.0=cuda
pytorch3d=0.7.7=py38_cu117_pyt200
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2rc1=pypi_0
qudida=0.0.4=pypi_0
questionary=1.10.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.32.3=py38h06a4308_0
requests-oauthlib=2.0.0=pypi_0
rich=13.7.1=pypi_0
rich-click=1.8.3=pypi_0
rlbench=1.2.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.10.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=2.12.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py38h06a4308_0
simple-knn=0.0.0=dev_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stable-diffusion-sdkit=2.1.3=pypi_0
sympy=1.12=py38h06a4308_0
sysroot_linux-64=2.17=h4a8ded7_16
tabulate=0.9.0=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pyhd8ed1ab_0
test-tube=0.7.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
timeout-decorator=0.5.0=pypi_0
timm=0.6.11=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torchaudio=2.0.0=py38_cu117
torchmetrics=0.6.0=pypi_0
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu117
tornado=6.4.1=pypi_0
tqdm=4.66.4=pyhd8ed1ab_0
transformers=4.26.1=pypi_0
trimesh=4.4.3=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=py38h06a4308_0
tzdata=2024a=h0c530f3_0
urllib3=1.26.19=pypi_0
visdom=0.2.4=pypi_0
wandb=0.14.0=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py38h06a4308_0
xformers=0.0.18=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
yarl=1.9.4=pypi_0
yarr=0.1=dev_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
